fix(score), feat(score): change max subject scores route, score datatype changed, top by limit added

+++ src/score/score_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Date, ForeignKey
+from sqlalchemy import Integer, Date, ForeignKey, Float
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from database.database import Base
 
@@ -9,7 +9,7 @@ class Score(Base):
     score_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     student_id: Mapped[int] = mapped_column(Integer, ForeignKey("student.student_id"), primary_key=True, autoincrement=False)
     subject_id: Mapped[int] = mapped_column(Integer, ForeignKey("subject.subject_id"), primary_key=True, autoincrement=False)
-    score: Mapped[int] = mapped_column(Integer)
+    score: Mapped[float] = mapped_column(Float)
     date: Mapped[str] = mapped_column(Date)
 
     student = relationship("Student", back_populates="scores")
